new: dev: test fuzzy-match for analyze script, consolidate code
* limit fuzzy match results using startswith input ID string
* cleanup package modules and tests, move initial profile funcs
