+++ tests/test_misc.py
@@ -3,13 +3,9 @@ from pathlib import Path
 import pytest
 from munch import Munch
 
-from ymltoxml.utils import (
-    StrYAML,
-    get_filelist,
-    get_profile_sets,
-    get_profile_type,
-    load_config,
-)
+from ymltoxml import utils
+from ymltoxml.utils import StrYAML, get_filelist, load_config
+
 
 def test_get_filelist():
     test_path = Path('docs') / 'source' / 'index.rst'
@@ -25,32 +21,6 @@ def test_get_filelist_debug():
     assert isinstance(files, list)
     assert len(files) == 6
     assert str(test_path) in files
-
-
-def test_get_profile_sets():
-    dirpath = 'tests/data/'
-    id_sets = get_profile_sets(dirpath)
-    assert len(id_sets) == 2
-    assert isinstance(id_sets, tuple)
-    for item in id_sets:
-        assert len(item) == 4
-    for item in id_sets[0]:
-        assert isinstance(item, set)
-    assert id_sets[0][0] > id_sets[0][1]
-
-
-def test_get_profile_type():
-    filename = 'PRIVACY-ids.txt'
-    profile_type = get_profile_type(filename)
-    assert isinstance(profile_type, str)
-    assert profile_type == 'PRIVACY'
-
-
-def test_get_profile_type_debug():
-    filename = 'PRIVACY-ids.txt'
-    profile_type = get_profile_type(filename, debug=True)
-    assert isinstance(profile_type, str)
-    assert profile_type == 'PRIVACY'
 
 
 def test_str_dumper():
@@ -84,3 +54,12 @@ def test_load_yasort_config():
     assert isinstance(popts, Munch)
     assert hasattr(popts, 'has_parent_key')
     assert pfile.stem == 'yasort' or '.yasort'
+
+
+def test_load_yagrep_config():
+    popts, pfile = load_config(yagrep=True)
+
+    assert isinstance(pfile, Path)
+    assert isinstance(popts, Munch)
+    assert hasattr(popts, 'default_separator')
+    assert pfile.stem == 'yagrep' or '.yagrep'
