+++ tox.ini
@@ -143,8 +143,9 @@ allowlist_externals =
 
 deps =
     {[base]deps}
-    pymavlink
-    fuzzyset2
+    # pymavlink
+    fuzzy-match
+    numpy
     -e .
 
 commands =
